server/requirements.txt
# Consider creating separate development/production requirements if the image bloats
asgiref==3.8.1
attrs==25.1.0
autopep8==2.3.2
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-silk==5.3.2
djangorestframework==3.15.2
djangorestframework-gis==1.1
drf-spectacular==0.28.0
gprof2dot==2024.6.6
gunicorn==23.0.0
idna==3.11
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
numpy==2.3.4
packaging==25.0
pandas==2.3.3
psycopg==3.2.3
pyarrow==22.0.0
pycodestyle==2.12.1
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.5
rpds-py==0.22.3
ruff==0.14.1
six==1.17.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2024.2
uritemplate==4.1.1
urllib3==2.5.0
